Refactor `power` and `boot` commands into `chassis`
Move `power` and `boot` packages into `chassis` as sub-packages. This makes more sense as they are both sub-commands of `chassis` anyway. Doing this removes the need to have `chassis.go` in both packages.
