+++ pkg/cmd/cli/chassis/power/cycle.go
