+++ pkg/cmd/cli/chassis/power/nmi.go
