+++ pkg/cmd/cli/chassis/power/off.go
