+++ pkg/cmd/cli/chassis/power/on.go
