+++ pkg/cmd/cli/chassis/power/reset.go
