+++ pkg/cmd/cli/chassis/power/status.go
