+++ pkg/cmd/cli/show/show.go
@@ -27,7 +27,7 @@
 package show
 
 import (
-	"github.com/Cray-HPE/gru/pkg/cmd/cli/boot"
+	"github.com/Cray-HPE/gru/pkg/cmd/cli/chassis/boot"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli/disks"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli/nics"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli/system"
